[clang-tidy] Fix readability-identifier-naming hungarian C type detection
In C, structs and enums variables have `struct` and `enum` keywords (for
example: `struct my_struct_t my_var;`). Fix the type parsing to correctly
use the struct/enum name instead of just `struct` or `enum`.

This way, the user can specify a UserDefinedType option with the name of a
C struct or enum with a custom prefix.